Install Jest and Enzyme, build CircleCI in github repo
