+++ .babelrc
@@ -1,0 +1,7 @@
+{
+  "presets": [
+    "react",
+    "env"
+  ],
+  "plugins": ["transform-react-jsx"],
+}
