+++ package.json
@@ -7,7 +7,7 @@
     "server": "nodemon ./server/index.js",
     "react-dev": "webpack -d --watch ",
     "seed-database": "node data/seed.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "jest"
   },
   "author": "",
   "license": "ISC",
@@ -16,7 +16,6 @@
     "babel-core": "^6.26.0",
     "babel-loader": "^7.1.2",
     "babel-preset-es2015": "^6.24.1",
-    "babel-preset-react": "^6.24.1",
     "body-parser": "^1.18.2",
     "express": "^4.16.2",
     "jquery": "^3.3.1",
@@ -37,12 +36,20 @@
   },
   "devDependencies": {
     "babel-eslint": "^8.2.2",
+    "babel-jest": "^22.4.1",
     "babel-preset-airbnb": "^2.4.0",
+    "babel-preset-env": "^1.6.1",
+    "babel-preset-react": "^6.24.1",
+    "enzyme": "^3.3.0",
+    "enzyme-adapter-react-16": "^1.1.1",
+    "enzyme-to-json": "^3.3.1",
     "eslint": "^4.18.1",
     "eslint-config-airbnb": "^16.1.0",
     "eslint-plugin-import": "^2.9.0",
     "eslint-plugin-jsx-a11y": "^6.0.3",
     "eslint-plugin-react": "^7.7.0",
-    "prettier-eslint": "^8.8.1"
+    "jest": "^22.4.2",
+    "prettier-eslint": "^8.8.1",
+    "react-test-renderer": "^16.2.0"
   }
 }
